Chatbot/requirements.txt
python-dotenv
certifi
requests
streamlit

langchain
langchain-community
langchain-ollama
faiss-cpu

groq==0.31.0
langchain_groq==0.3.7

pypdf
tiktoken

streamlit-webrtc
pydub
